push BE Bikin insertheaderdetailtran check di fe

+++ internal/entity/skirpsionline-BE/td_transaction.go
@@ -14,8 +14,8 @@ type TD_Transaction2 struct {
 	TraId             string `db:"tra_id" json:"tra_id"`
 	ProdId             string    `db:"prod_id" json:"prod_id"`
 	TraDtlQty         int    `db:"tradtl_qty" json:"tradtl_qty"`
-	TraDtlPrice			int		`db:"tradtl_price" json:"tradtl_price"`
-	TraDtlAmount		int		`db:"tradtl_amount" json:"tradtl_amount"`
+	TraDtlPrice			float64		`db:"tradtl_price" json:"tradtl_price"`
+	TraDtlAmount		float64		`db:"tradtl_amount" json:"tradtl_amount"`
 }
 
 type InsertDetailTransaction struct {
